+++ components/types.ts
@@ -1,0 +1,13 @@
+export type CurrentDateType = {
+  day: string,
+  date: string
+}
+
+export type CalendarEventsType = {
+  id: number, 
+  title: string,
+  date: Date, 
+  day: number,
+  startDate: Date,
+  startTime: Date
+}
